python/jetstream/ocr/requirements.txt
nats-py
pytesseract
Pillow
